Add time-related columns

Adds useful columns for subsequent time calculations and filtering.

compute_timefilter_facet_at(data = NULL, timebin = NULL, facet = NULL)

compute_timefilter_facet(..., timebin, facet)

compute_timefilter_at(data = NULL, timebin = NULL, facet = NULL)

compute_timefilter(..., timebin, facet)

Arguments

data
data.frame.

timebin
bare for NSE; character for SE. Name of column in data to use for time axis.

facet
bare for NSE; character for SE. Name of column in data used for facetting.

...
dots. Parameters to pass directly to visualize_time().

Value

list.

list.

Details

Calls compute_timefilter_facet_at(). (Adds a dummy facet column if none is specified.) Adds columns for years, months, days, and hours elapsed. (Calls in compute_time_elapsed() facetple times.) Returns list, where data is augmented data.frame, date_start is the very LAST Date-time value, and date_end is the very FIRST Date-time value. The date_start and date_end values are defined in this manner such that all data for each facet variable fits in a singular frame (i.e. there are not 'gaps' where data exists for one facet variable and not all others). This determination assumes that the data is continuous.