Prepare data for text analysis.

Create a tidy data.frame of unigrams.

tidify_to_bigrams_at(data = NULL, text = NULL, rgx_pattern, rgx_replacement,
  stopwords = TRUE, stopwords_lexicon, rgx_ignore_custom)

tidify_to_bigrams(..., text)

Arguments

data
data.frame.

text
bare for NSE; character for SE. Name of column in data to parse.

rgx_pattern
character. Regular expression to substitute.

rgx_replacement
character. Regular expression used as replacement.

stopwords
logical. Whether or not to remove stopwords. Default: TRUE

stopwords_lexicon
character. If not specified, then uses all stop words available in tidytext package. Otherwise, must be a lexicon available in the tidytext package.

rgx_ignore_custom
character. Custom regular expression to remove.

...
dots. Parameters to pass directly to visualize_time().

Value

data.frame.

Details

Heavily influenced by Text Mining with R. Creates output columns 'word', 'word1', and 'word2'. ('word' is simply 'word1' and 'word2' toegether.)