Visualize counts

Visualize token counts with a token cloud.

visualize_cnts_wordcloud_at(data = NULL, token = NULL,
  wordcloud_base = default_wordcloud(data, token),
  wordcloud_params = list())

visualize_cnts_wordcloud(..., token)

visualize_cnts_wordcloud_facet_at(data = NULL, ..., facet = NULL,
  value_facet = NULL)

visualize_cnts_wordcloud_facet(..., token, facet)

Arguments

data
data.frame.

token
bare for NSE; character for SE. Name of column in data to use for words argument in wordcloud::wordcloud().

wordcloud_base
function. Works similarly to other _base arguments.

wordcloud_params
function. Works similarly to other _params arguments.

...
dots. Parameters to pass to visualize_cnts_wordcloud_at().

facet
bare for NSE; character for SE. Name of column in data to use to create a separate wordcloud.

value_facet
charater or numeric. Value by which to filter the column specified by facet. Ignored if all = TRUE. Defaults are provided internally based on value of all and number of facet values.

Value

plot.

plots.

Details

facet version is for creating multiple plots simulatenously (probably using a purrr 'map' function). Calls wordcloud::wordcloud().