Visualize correlations

Visualize correlations with a network

visualize_corrs_network_at(..., resize_points = TRUE, size_point = "n",
  add_point_labels = TRUE, color_point = "grey50", shape_point = 21,
  seed = 42, labs_base = default_labs(), labs_params = list(title =
  "Network of Word Pairwise Correlations"), theme_base = default_theme(void =
  TRUE), theme_params = list())

visualize_corrs_network(..., token, feature)

Arguments

...
dots. Parameters passed directly to compute_corrs().

resize_points
logical. Indicates whether or not to make size of points correspond to count of words.

size_point
bare for NSE; character for SE. Name of column in data to use for point sizing.

add_point_labels
logical. Indicates whether or not to add labels to points.

color_point
character. Hex value of color_value for points.

shape_point
numeric.

seed
numeric. Used to initialize ggraph::ggraph().

labs_base
ggplot2::labs() function. Defaults to a pre-determined set of values. It is recommended NOT to modify this argument; isntead, the _params argument should be used for customization.

labs_params
list. Additional parameters to pass to ggplot2::labs() to use in addition to and/or override the parameters set in the _base parameter.

theme_base
ggplot2::theme() function. (e.g. as ggplot2::theme_minimal().) A custom theme is supplied as a default. As with the the labs_base argument, it is NOT recommended to change this directly.

theme_params
list. Additional parameters to pass to ggplot2::labs(). Should be used in the same manner as labs_params for customization.

token
bare for NSE; character for SE. Name of column in data to use as unit for counting. Probably something like 'word' or 'bigram'.

feature
bare for NSE; character for SE. Name of column in data to use as feature in widyr::pairwise_cor().

Value

gg.

Details

It's not straightforward to create _base and _params arguments for this function to map to the ggraph arguments, so explicit arguments are used (i.e. color_point shape_point

See also

https://www.tidytextmining.com/ngrams.html. http://varianceexplained.org/r/seven-fav-packages/.