Visualize hourly data

Visualize time-based data aggregated over hours.

visualize_time_hh_at(data = NULL, timebin = NULL, facet = NULL,
  color = facet, scale_manual_base = default_scale_manual(),
  scale_manual_params = list(), labs_base = default_labs(),
  labs_params = list(title = "Count by Hour of Day"),
  theme_base = default_theme(), theme_params = list(panel.grid =
  ggplot2::element_blank()))

visualize_time_hh(..., timebin, color, facet)

Arguments

data
data.frame.

timebin
bare for NSE; character for SE. Name of column in data to use for time. Probably likely something like 'hh' or 'hour'.

facet
bare for NSE; character for SE. Name of column in data to use one of main axes, not for facetting.

color
bare for NSE; character for SE. Name of column in data to use for color basis. Even if fill is the actual ggplot2 aesthetic, the internal code will 'coerce' this column to fill. In order to simplify internal code, MUST be specified.

scale_manual_base
list. Parameters to pass to ggplot2::scale_fill_manual() or ggplot2::scale_color_manual(), depending on the aesthetic mapping. A default values argument is specified in the internal function default_scale_manual(), so if overwriting explicitly, then values should be included in the list.

scale_manual_params
list. Additional parameters to pass to ggplot2::scale_fill_manual() or ggplot2::scale_color_manual() (e.g. breaks, labels, or name). The default values specified by scale_manual_base = default_scale_manual() can be overwritten a values argument in this list (or directly with a values argument in the scale_manual_base list.

labs_base
ggplot2::labs() function. Defaults to a pre-determined set of values. It is recommended NOT to modify this argument; isntead, the _params argument should be used for customization.

labs_params
list. Additional parameters to pass to ggplot2::labs() to use in addition to and/or override the parameters set in the _base parameter.

theme_base
ggplot2::theme() function. (e.g. as ggplot2::theme_minimal().) A custom theme is supplied as a default. As with the the labs_base argument, it is NOT recommended to change this directly.

theme_params
list. Additional parameters to pass to ggplot2::labs(). Should be used in the same manner as labs_params for customization.

...
dots. Parameters to pass directly to visualize_time().

Details

This function is 'unique' because it calls ggplot2::geom_violin(). Also, it does not actually use facet for facetting.